Refactor ROM menu architecture
