+++ platform/pico2w/src/display/mod.rs
@@ -14,6 +14,15 @@ pub mod fb;
 pub mod font;
 #[cfg(target_arch = "arm")]
 pub mod hw;
+mod loading;
+mod menu;
+mod text;
+
+pub use loading::{loading_bar_y_range, render_loading_row};
+pub use menu::{
+    menu_item_needs_marquee, menu_item_text_window, menu_item_y_range, render_menu_row,
+    MENU_MARQUEE_REDRAW_FRAMES,
+};
 
 use embedded_graphics::draw_target::DrawTarget;
 use embedded_graphics::geometry::{Dimensions, Point, Size};
@@ -338,451 +347,6 @@ mod logo_font {
 }
 
 // ---------------------------------------------------------------------------
-// Menu rendering
-// ---------------------------------------------------------------------------
-
-use crate::menu::MenuFrame;
-
-// DMG palette in big-endian RGB565 — the byte order the ILI9341 SPI bus expects.
-const C0_BE: [u8; 2] = [0xE7, 0xDA]; // #E0F8D0
-const C1_BE: [u8; 2] = [0x8E, 0x0E]; // #88C070
-const C2_BE: [u8; 2] = [0x33, 0x4A]; // #346856
-const C3_BE: [u8; 2] = [0x08, 0xC4]; // #081820
-
-// Layout constants (pixels on the 240×320 display).
-const HEADER_H: u16 = 40;
-const SEPARATOR_H: u16 = 4;
-const ITEMS_START_Y: u16 = HEADER_H + SEPARATOR_H;
-const ITEM_H: u16 = 32;
-const ITEM_TEXT_PAD: u16 = 8; // top padding within each item slot
-const FOOTER_SEP_Y: u16 = 280;
-const FOOTER_Y: u16 = FOOTER_SEP_Y + SEPARATOR_H;
-const CURSOR_X: usize = 10;
-const ITEM_TEXT_X: usize = 28;
-const ITEM_TEXT_RIGHT_X: usize = 236;
-const ITEM_MARKER_GAP: usize = 8;
-const MARKER_X: usize = 196; // right-side "loaded" indicator column
-const CHAR_W: usize = 8; // glyph width in pixels
-const SCALE: usize = 2; // render scale for items and title
-const MARQUEE_HOLD_FRAMES: u32 = 30;
-const MARQUEE_GAP_CHARS: usize = 3;
-const MARQUEE_FRAMES_PER_STEP: u32 = 1;
-
-pub const MENU_MARQUEE_REDRAW_FRAMES: u32 = MARQUEE_FRAMES_PER_STEP;
-
-/// Return true when a menu item needs marquee animation in the ROM list.
-pub fn menu_item_needs_marquee(text: &str, marked: bool) -> bool {
-    text_screen_width(text.as_bytes(), SCALE) > menu_item_text_window_width(marked)
-}
-
-/// Return the inclusive-start/exclusive-end y-range for a menu item slot.
-pub fn menu_item_y_range(slot: usize) -> Option<(u16, u16)> {
-    let y_start = ITEMS_START_Y.checked_add(slot.checked_mul(ITEM_H as usize)? as u16)?;
-    if y_start >= FOOTER_SEP_Y {
-        return None;
-    }
-    Some((y_start, (y_start + ITEM_H).min(FOOTER_SEP_Y)))
-}
-
-/// Return the text-only window for a menu item slot.
-pub fn menu_item_text_window(slot: usize, marked: bool) -> Option<(u16, u16, u16, u16)> {
-    let (item_y_start, item_y_end) = menu_item_y_range(slot)?;
-    let y_start = item_y_start + ITEM_TEXT_PAD;
-    let y_end = (y_start + (8 * SCALE) as u16).min(item_y_end);
-    let x_start = ITEM_TEXT_X as u16;
-    let x_end = menu_item_text_window_end(marked) as u16;
-    if x_end <= x_start || y_end <= y_start {
-        return None;
-    }
-    Some((x_start, x_end, y_start, y_end))
-}
-
-/// Render one horizontal scan-line of a menu into `row` (240 pixels × 2 bytes BE).
-///
-/// Call for every `y` in `0..320` to produce a complete menu frame.
-/// The buffer is entirely overwritten each call.
-pub fn render_menu_row(frame: &MenuFrame<'_>, y: u16, row: &mut [u8; 480]) {
-    let in_top_sep = y >= HEADER_H && y < HEADER_H + SEPARATOR_H;
-    let in_footer_sep = y >= FOOTER_SEP_Y && y < FOOTER_Y;
-    let in_items = y >= ITEMS_START_Y && y < FOOTER_SEP_Y;
-
-    fill_row(
-        row,
-        if in_top_sep || in_footer_sep {
-            C2_BE
-        } else {
-            C3_BE
-        },
-    );
-
-    if y < HEADER_H {
-        render_menu_header_row(frame.title, y, row);
-    }
-    if in_items {
-        render_menu_items_row(frame, y, row);
-    }
-    if y >= FOOTER_Y {
-        render_menu_footer_row(y, row);
-    }
-}
-
-fn render_menu_header_row(title: &str, y: u16, row: &mut [u8; 480]) {
-    let title_screen_h = (8 * SCALE) as u16;
-    let title_top = (HEADER_H - title_screen_h) / 2;
-    if y < title_top || y >= title_top + title_screen_h {
-        return;
-    }
-    let glyph_row = ((y - title_top) as usize) / SCALE;
-    let title_screen_w = (title.len() * CHAR_W * SCALE) as usize;
-    let title_x = (SCREEN_W as usize).saturating_sub(title_screen_w) / 2;
-    write_text_row(
-        row,
-        title.as_bytes(),
-        title_x,
-        glyph_row,
-        SCALE,
-        C0_BE,
-        C3_BE,
-    );
-}
-
-fn render_menu_items_row(frame: &MenuFrame<'_>, y: u16, row: &mut [u8; 480]) {
-    let slot = ((y - ITEMS_START_Y) / ITEM_H) as usize;
-    if slot >= frame.items.len() {
-        return;
-    }
-    let slot_top = ITEMS_START_Y + slot as u16 * ITEM_H;
-    let selected = slot == frame.selected;
-    let enabled = frame.enabled.get(slot).copied().unwrap_or(true);
-
-    fill_row(row, if selected { C2_BE } else { C3_BE });
-
-    let text_top = slot_top + ITEM_TEXT_PAD;
-    let text_bottom = text_top + (8 * SCALE) as u16;
-    if y < text_top || y >= text_bottom {
-        return;
-    }
-
-    let glyph_row = ((y - text_top) as usize) / SCALE;
-    let text_color = if !enabled {
-        C2_BE
-    } else if selected {
-        C0_BE
-    } else {
-        C1_BE
-    };
-    let item_bg = if selected { C2_BE } else { C3_BE };
-
-    let cursor: &[u8] = if selected { b">" } else { b" " };
-    write_text_row(row, cursor, CURSOR_X, glyph_row, SCALE, text_color, item_bg);
-    let text = frame.items[slot].as_bytes();
-    let text_right = menu_item_text_window_end(frame.marked == Some(slot));
-    if selected && text_screen_width(text, SCALE) > text_right.saturating_sub(ITEM_TEXT_X) {
-        write_marquee_text_row(
-            row,
-            text,
-            ITEM_TEXT_X,
-            text_right,
-            glyph_row,
-            SCALE,
-            frame.marquee_frame,
-            text_color,
-            item_bg,
-        );
-    } else {
-        write_truncated_text_row(
-            row,
-            text,
-            ITEM_TEXT_X,
-            text_right,
-            glyph_row,
-            SCALE,
-            text_color,
-            item_bg,
-        );
-    }
-    if frame.marked == Some(slot) {
-        write_text_row(row, b"*", MARKER_X, glyph_row, SCALE, C0_BE, item_bg);
-    }
-}
-
-fn menu_item_text_window_end(marked: bool) -> usize {
-    if marked {
-        MARKER_X.saturating_sub(ITEM_MARKER_GAP)
-    } else {
-        ITEM_TEXT_RIGHT_X
-    }
-}
-
-fn menu_item_text_window_width(marked: bool) -> usize {
-    menu_item_text_window_end(marked).saturating_sub(ITEM_TEXT_X)
-}
-
-fn render_menu_footer_row(y: u16, row: &mut [u8; 480]) {
-    const FOOTER: &[u8] = b"A:SELECT  B:BACK";
-    let footer_text_h = 8u16;
-    let footer_top = FOOTER_Y + (SCREEN_H as u16 - FOOTER_Y - footer_text_h) / 2;
-    if y < footer_top || y >= footer_top + footer_text_h {
-        return;
-    }
-    let glyph_row = (y - footer_top) as usize;
-    let footer_w = FOOTER.len() * CHAR_W;
-    let footer_x = (SCREEN_W as usize).saturating_sub(footer_w) / 2;
-    write_text_row(row, FOOTER, footer_x, glyph_row, 1, C1_BE, C3_BE);
-}
-
-// ---------------------------------------------------------------------------
-// Loading screen rendering
-// ---------------------------------------------------------------------------
-
-const LOADING_FILENAME_Y: u16 = 140;
-const LOADING_BAR_TOP: u16 = 200;
-const LOADING_BAR_BOTTOM: u16 = 220;
-const LOADING_BAR_X0: usize = 20;
-const LOADING_BAR_X1: usize = 220;
-const LOADING_FILENAME_X0: usize = 8;
-const LOADING_FILENAME_X1: usize = 232;
-
-/// Return the inclusive-start/exclusive-end y-range for the loading progress bar.
-pub fn loading_bar_y_range() -> (u16, u16) {
-    (LOADING_BAR_TOP, LOADING_BAR_BOTTOM)
-}
-
-/// Render one scan-line of the ROM loading screen.
-///
-/// `banks_done` / `total_banks` drives the progress bar fill (0–100%).
-/// Call for every `y` in `0..320` to produce a complete frame.
-pub fn render_loading_row(
-    filename: &str,
-    banks_done: u32,
-    total_banks: u32,
-    _marquee_frame: u32,
-    y: u16,
-    row: &mut [u8; 480],
-) {
-    let in_top_sep = y >= HEADER_H && y < HEADER_H + SEPARATOR_H;
-    fill_row(row, if in_top_sep { C2_BE } else { C3_BE });
-
-    if y < HEADER_H {
-        render_loading_header_row(y, row);
-    }
-    if y >= LOADING_FILENAME_Y && y < LOADING_FILENAME_Y + 8 {
-        render_loading_filename_row(filename, y, row);
-    }
-    if y >= LOADING_BAR_TOP && y < LOADING_BAR_BOTTOM {
-        render_loading_bar_row(banks_done, total_banks, row);
-    }
-}
-
-fn render_loading_header_row(y: u16, row: &mut [u8; 480]) {
-    let title_screen_h = (8 * SCALE) as u16;
-    let title_top = (HEADER_H - title_screen_h) / 2;
-    if y < title_top || y >= title_top + title_screen_h {
-        return;
-    }
-    let glyph_row = ((y - title_top) as usize) / SCALE;
-    let title = b"LOADING";
-    let title_w = title.len() * CHAR_W * SCALE;
-    let title_x = (SCREEN_W as usize).saturating_sub(title_w) / 2;
-    write_text_row(row, title, title_x, glyph_row, SCALE, C0_BE, C3_BE);
-}
-
-fn render_loading_filename_row(filename: &str, y: u16, row: &mut [u8; 480]) {
-    let glyph_row = (y - LOADING_FILENAME_Y) as usize;
-    let text = filename.as_bytes();
-    let text_w = text_screen_width(text, 1);
-    let window_w = LOADING_FILENAME_X1.saturating_sub(LOADING_FILENAME_X0);
-    if text_w > window_w {
-        write_truncated_text_row(
-            row,
-            text,
-            LOADING_FILENAME_X0,
-            LOADING_FILENAME_X1,
-            glyph_row,
-            1,
-            C1_BE,
-            C3_BE,
-        );
-    } else {
-        let text_x = (SCREEN_W as usize).saturating_sub(text_w) / 2;
-        write_text_row(row, text, text_x, glyph_row, 1, C1_BE, C3_BE);
-    }
-}
-
-fn render_loading_bar_row(banks_done: u32, total_banks: u32, row: &mut [u8; 480]) {
-    let bar_w = LOADING_BAR_X1 - LOADING_BAR_X0;
-    let filled = if total_banks > 0 {
-        (bar_w as u64 * banks_done as u64 / total_banks as u64) as usize
-    } else {
-        0
-    };
-    for px in LOADING_BAR_X0..LOADING_BAR_X1 {
-        row[px * 2] = C2_BE[0];
-        row[px * 2 + 1] = C2_BE[1];
-    }
-    for px in LOADING_BAR_X0..LOADING_BAR_X0 + filled.min(bar_w) {
-        row[px * 2] = C0_BE[0];
-        row[px * 2 + 1] = C0_BE[1];
-    }
-}
-
-/// Fill every pixel in `row` with `color` (big-endian RGB565).
-fn fill_row(row: &mut [u8; 480], color: [u8; 2]) {
-    let mut i = 0;
-    while i < 480 {
-        row[i] = color[0];
-        row[i + 1] = color[1];
-        i += 2;
-    }
-}
-
-/// Write `text` into `row` at horizontal pixel offset `x_start`.
-///
-/// `glyph_row` selects which of the 8 bitmap rows to render.
-/// `scale` is 1 or 2 (each source pixel becomes `scale` screen pixels wide).
-fn write_text_row(
-    row: &mut [u8; 480],
-    text: &[u8],
-    x_start: usize,
-    glyph_row: usize,
-    scale: usize,
-    fg: [u8; 2],
-    bg: [u8; 2],
-) {
-    let char_screen_w = CHAR_W * scale;
-    for (char_idx, &ch) in text.iter().enumerate() {
-        let bitmap = font::glyph_row(ch, glyph_row);
-        let char_x = x_start + char_idx * char_screen_w;
-        for glyph_x in 0..CHAR_W {
-            // Font bitmaps use bit 0 = leftmost pixel (LSB-first convention).
-            let lit = (bitmap >> glyph_x) & 1 != 0;
-            let color = if lit { fg } else { bg };
-            for sx in 0..scale {
-                let px = char_x + glyph_x * scale + sx;
-                if px < SCREEN_W as usize {
-                    row[px * 2] = color[0];
-                    row[px * 2 + 1] = color[1];
-                }
-            }
-        }
-    }
-}
-
-fn write_marquee_text_row(
-    row: &mut [u8; 480],
-    text: &[u8],
-    x_start: usize,
-    x_end: usize,
-    glyph_row: usize,
-    scale: usize,
-    marquee_frame: u32,
-    fg: [u8; 2],
-    bg: [u8; 2],
-) {
-    let scroll_px = marquee_scroll_px(text, scale, marquee_frame);
-    write_text_window_row(
-        row, text, x_start, x_end, glyph_row, scale, scroll_px, true, fg, bg,
-    );
-}
-
-fn write_truncated_text_row(
-    row: &mut [u8; 480],
-    text: &[u8],
-    x_start: usize,
-    x_end: usize,
-    glyph_row: usize,
-    scale: usize,
-    fg: [u8; 2],
-    bg: [u8; 2],
-) {
-    let char_screen_w = CHAR_W * scale;
-    if char_screen_w == 0 {
-        return;
-    }
-
-    let max_chars = x_end.saturating_sub(x_start) / char_screen_w;
-    let visible = text.len().min(max_chars);
-    write_text_row(row, &text[..visible], x_start, glyph_row, scale, fg, bg);
-}
-
-fn write_text_window_row(
-    row: &mut [u8; 480],
-    text: &[u8],
-    x_start: usize,
-    x_end: usize,
-    glyph_row: usize,
-    scale: usize,
-    scroll_px: usize,
-    wrap: bool,
-    fg: [u8; 2],
-    bg: [u8; 2],
-) {
-    let x0 = x_start.min(SCREEN_W as usize);
-    let x1 = x_end.min(SCREEN_W as usize);
-    if x1 <= x0 || text.is_empty() || scale == 0 {
-        return;
-    }
-
-    let text_w = text_screen_width(text, scale);
-    if text_w == 0 {
-        return;
-    }
-    let gap_w = MARQUEE_GAP_CHARS * CHAR_W * scale;
-    let period_w = text_w + if wrap { gap_w } else { 0 };
-
-    for px in x0..x1 {
-        let source_px = px - x0 + scroll_px;
-        let source_px = if wrap {
-            source_px % period_w
-        } else {
-            source_px
-        };
-        let color = if source_px < text_w && text_pixel_lit(text, glyph_row, scale, source_px) {
-            fg
-        } else {
-            bg
-        };
-        row[px * 2] = color[0];
-        row[px * 2 + 1] = color[1];
-    }
-}
-
-fn marquee_scroll_px(text: &[u8], scale: usize, marquee_frame: u32) -> usize {
-    let text_w = text_screen_width(text, scale);
-    let period_w = text_w + MARQUEE_GAP_CHARS * CHAR_W * scale;
-    if period_w == 0 {
-        return 0;
-    }
-
-    let step_px = scale.max(1);
-    let scroll_steps = period_w.div_ceil(step_px);
-    let scroll_frames = (scroll_steps as u32).saturating_mul(MARQUEE_FRAMES_PER_STEP);
-    let phase = marquee_frame % (MARQUEE_HOLD_FRAMES + scroll_frames);
-    if phase < MARQUEE_HOLD_FRAMES {
-        0
-    } else {
-        let step = ((phase - MARQUEE_HOLD_FRAMES) / MARQUEE_FRAMES_PER_STEP) as usize;
-        (step * step_px) % period_w
-    }
-}
-
-fn text_pixel_lit(text: &[u8], glyph_row: usize, scale: usize, source_px: usize) -> bool {
-    let char_screen_w = CHAR_W * scale;
-    let char_idx = source_px / char_screen_w;
-    let Some(&ch) = text.get(char_idx) else {
-        return false;
-    };
-    let glyph_x = (source_px % char_screen_w) / scale;
-    let bitmap = font::glyph_row(ch, glyph_row);
-    // Keep the existing LSB-first convention used by write_text_row.
-    (bitmap >> glyph_x) & 1 != 0
-}
-
-fn text_screen_width(text: &[u8], scale: usize) -> usize {
-    text.len() * CHAR_W * scale
-}
-
-// ---------------------------------------------------------------------------
 // Tests
 // ---------------------------------------------------------------------------
 
@@ -794,54 +358,6 @@ mod tests {
 
     fn make_display() -> Display<FbDisplay> {
         Display::from_draw_target(FbDisplay::new(SCREEN_W as u32, SCREEN_H as u32))
-    }
-
-    #[test]
-    fn menu_item_marquee_threshold_reserves_marker_space() {
-        assert!(!menu_item_needs_marquee("1234567890123", false));
-        assert!(menu_item_needs_marquee("12345678901234", false));
-        assert!(!menu_item_needs_marquee("1234567890", true));
-        assert!(menu_item_needs_marquee("12345678901", true));
-    }
-
-    #[test]
-    fn marquee_scroll_holds_before_moving() {
-        assert_eq!(marquee_scroll_px(b"LONG-ROM-NAME.GB", SCALE, 0), 0);
-        assert_eq!(
-            marquee_scroll_px(b"LONG-ROM-NAME.GB", SCALE, MARQUEE_HOLD_FRAMES - 1),
-            0
-        );
-        assert_eq!(
-            marquee_scroll_px(
-                b"LONG-ROM-NAME.GB",
-                SCALE,
-                MARQUEE_HOLD_FRAMES + MARQUEE_FRAMES_PER_STEP
-            ),
-            SCALE
-        );
-    }
-
-    #[test]
-    fn marquee_scroll_uses_slowed_full_period_before_wrapping() {
-        let text = b"LONG-ROM-NAME.GB";
-        let period_w = text_screen_width(text, SCALE) + MARQUEE_GAP_CHARS * CHAR_W * SCALE;
-        assert_eq!(
-            marquee_scroll_px(
-                text,
-                SCALE,
-                MARQUEE_HOLD_FRAMES
-                    + (period_w as u32 / SCALE as u32 - 1) * MARQUEE_FRAMES_PER_STEP
-            ),
-            period_w - SCALE
-        );
-        assert_eq!(
-            marquee_scroll_px(
-                text,
-                SCALE,
-                MARQUEE_HOLD_FRAMES + period_w as u32 / SCALE as u32 * MARQUEE_FRAMES_PER_STEP
-            ),
-            0
-        );
     }
 
     #[test]
