+++ platform/pico2w/src/menu.rs
@@ -56,6 +56,18 @@ pub enum RomListEffect {
     Back,
 }
 
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub enum RomPageSelection {
+    First,
+    Last,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct RomPageRequest {
+    pub offset: usize,
+    pub selection: RomPageSelection,
+}
+
 // ---------------------------------------------------------------------------
 // Render descriptor
 // ---------------------------------------------------------------------------
@@ -278,6 +290,38 @@ impl RomListLogic {
     /// Move selection to the final item on the current page.
     pub fn select_last(&mut self) {
         self.selected = self.page_len.saturating_sub(1);
+    }
+}
+
+pub fn rom_page_request(
+    effect: RomListEffect,
+    page_offset: usize,
+    page_size: usize,
+    total_roms: usize,
+    has_next: bool,
+) -> Option<RomPageRequest> {
+    if page_size == 0 || total_roms == 0 {
+        return None;
+    }
+
+    match effect {
+        RomListEffect::NextPage if has_next => Some(RomPageRequest {
+            offset: page_offset + page_size,
+            selection: RomPageSelection::First,
+        }),
+        RomListEffect::NextPage => Some(RomPageRequest {
+            offset: 0,
+            selection: RomPageSelection::First,
+        }),
+        RomListEffect::PrevPage if page_offset > 0 => Some(RomPageRequest {
+            offset: page_offset.saturating_sub(page_size),
+            selection: RomPageSelection::Last,
+        }),
+        RomListEffect::PrevPage => Some(RomPageRequest {
+            offset: (total_roms.saturating_sub(1) / page_size) * page_size,
+            selection: RomPageSelection::Last,
+        }),
+        _ => None,
     }
 }
 
@@ -540,5 +584,45 @@ mod tests {
     fn romlist_no_input_returns_none() {
         let mut logic = RomListLogic::new(3);
         assert_eq!(logic.handle(no_input()), RomListEffect::None);
+    }
+
+    #[test]
+    fn rom_page_request_down_moves_to_next_page_top() {
+        assert_eq!(
+            rom_page_request(RomListEffect::NextPage, 0, 7, 12, true),
+            Some(RomPageRequest {
+                offset: 7,
+                selection: RomPageSelection::First,
+            })
+        );
+    }
+
+    #[test]
+    fn rom_page_request_up_moves_to_previous_page_bottom() {
+        assert_eq!(
+            rom_page_request(RomListEffect::PrevPage, 7, 7, 12, true),
+            Some(RomPageRequest {
+                offset: 0,
+                selection: RomPageSelection::Last,
+            })
+        );
+    }
+
+    #[test]
+    fn rom_page_request_wraps_ends() {
+        assert_eq!(
+            rom_page_request(RomListEffect::NextPage, 7, 7, 12, false),
+            Some(RomPageRequest {
+                offset: 0,
+                selection: RomPageSelection::First,
+            })
+        );
+        assert_eq!(
+            rom_page_request(RomListEffect::PrevPage, 0, 7, 12, true),
+            Some(RomPageRequest {
+                offset: 7,
+                selection: RomPageSelection::Last,
+            })
+        );
     }
 }
